interger por long

+++ src/main/java/Uegg/appInmobiliaria/controladores/UsuarioControlador.java
@@ -37,7 +37,7 @@ public class UsuarioControlador {
             @RequestParam(required = false) Long dni,
             @RequestParam String direccion,
             @RequestParam(required = false) Integer codigoPostal,
-            @RequestParam(required = false) Integer telefono,
+            @RequestParam(required = false) Long telefono,
             @RequestParam String email,
             @RequestParam String pass,
             @RequestParam String pass2,
@@ -76,7 +76,7 @@ public class UsuarioControlador {
             @RequestParam(required = false) Long cuit,
             @RequestParam String direccion,
             @RequestParam(required = false) Integer codigoPostal,
-            @RequestParam(required = false) Integer telefono,
+            @RequestParam(required = false) Long telefono,
             @RequestParam String email,
             @RequestParam String pass,
             @RequestParam String pass2,
@@ -118,7 +118,7 @@ public class UsuarioControlador {
     @PostMapping("/modificar/{id}")
     public String modificar(
             @PathVariable String id,
-            @RequestParam Integer telefono,
+            @RequestParam Long telefono,
             @RequestParam String direccion,
             @RequestParam Integer codigoPostal,
             @RequestParam String email,
